adjusting to make the repo work self-contained
