+++ assets/app/models/data/pickles/nb-predictor.pkl
