+++ assets/requirements.txt
@@ -1,0 +1,64 @@
+altgraph==0.17.3
+anyio==3.6.2
+auto-py-to-exe==2.24.1
+blis==0.7.9
+bottle==0.12.23
+bottle-websocket==0.2.9
+catalogue==2.0.8
+certifi==2022.9.24
+cffi==1.15.1
+charset-normalizer==2.1.1
+click==8.1.3
+colorama==0.4.6
+confection==0.0.3
+cymem==2.0.7
+Eel==0.14.0
+en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
+fastapi==0.87.0
+future==0.18.2
+gevent==22.10.2
+gevent-websocket==0.10.1
+greenlet==2.0.1
+h11==0.14.0
+idna==3.4
+Jinja2==3.1.2
+joblib==1.2.0
+langcodes==3.3.0
+MarkupSafe==2.1.1
+murmurhash==1.0.9
+numpy==1.23.4
+packaging==21.3
+pandas==1.5.1
+pathy==0.8.1
+pefile==2022.5.30
+preshed==3.0.8
+pycparser==2.21
+pydantic==1.10.2
+pyinstaller==5.6.2
+pyinstaller-hooks-contrib==2022.13
+pyparsing==3.0.9
+python-dateutil==2.8.2
+pytz==2022.6
+pywin32-ctypes==0.2.0
+requests==2.28.1
+scikit-learn==0.24.2
+scipy==1.9.3
+six==1.16.0
+smart-open==5.2.1
+sniffio==1.3.0
+spacy==3.4.3
+spacy-legacy==3.0.10
+spacy-loggers==1.0.3
+srsly==2.4.5
+starlette==0.21.0
+thinc==8.1.5
+threadpoolctl==3.1.0
+tqdm==4.64.1
+typer==0.7.0
+typing_extensions==4.4.0
+urllib3==1.26.12
+uvicorn==0.19.0
+wasabi==0.10.1
+whichcraft==0.6.1
+zope.event==4.5.0
+zope.interface==5.5.2
